fix: 评审整改（移除EF迁移/明文密码/补齐认证/重构超管端/审计日志修正）
- 移除 backend/Migrations/ 及 Microsoft.EntityFrameworkCore.Design 包
- 从 appsettings.Development.json 移除明文密码，改用 user-secrets
- 补齐认证接口（AUTH-01~05 + CAPTCHA-01 + PWD-01）
- 重构超管端控制器，按契约路径创建 Colleges/Majors/Students/Admins/AuditEvents/Reports/Violations 控制器（返回 501）
- 修正审计日志实体与服务，对齐 D_Audit_Event 表结构
- 删除死代码 Models/Enums/UserRole.cs

+++ backend/DTO/RepairDtos.cs
@@ -13,7 +13,7 @@ public sealed class SubmitRepairTicketRequest
     [RegularExpression("^(普通|紧急)$", ErrorMessage = "紧急程度只能为普通或紧急")]
     public string Urgency { get; set; } = string.Empty;
 
-    public long? RoomId { get; set; }
+    public int? RoomId { get; set; }
 }
 
 public sealed class RepairTicketQueryDto
@@ -36,7 +36,7 @@ public sealed class RepairTicketDto
 {
     public long TicketId { get; set; }
     public string StudentId { get; set; } = string.Empty;
-    public long RoomId { get; set; }
+    public int RoomId { get; set; }
     public string Description { get; set; } = string.Empty;
     public DateTime SubmitTime { get; set; }
     public string Status { get; set; } = string.Empty;
